chore: change name to AccountService
