+++ account-service/.editorconfig
