+++ account-service/src/main/java/com/account/service/HelloService.java
@@ -1,0 +1,11 @@
+package com.account.service;
+
+import com.account.vo.HelloRequestVO;
+
+public interface HelloService {
+
+    String echo(String input);
+
+    String hello(HelloRequestVO helloRequestVO);
+}
+
