+++ account-service/src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,0 +1,8 @@
+databaseChangeLog:
+- changeSet:
+      id: 'user-service-1'
+      author: Emberow
+      changes:
+        - sqlFile:
+            encoding: utf8
+            path: classpath:db/changelog/sqlfile/20240610_1.sql
