+++ account-service/src/test/java/com/account/AccountServiceTests.java
@@ -1,0 +1,13 @@
+package com.account;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiGatewayTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
