--- user-service/src/main/java/com/userservice/UserService.java
@@ -1,18 +1,0 @@
-package com.userservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.metrics.buffering.BufferingApplicationStartup;
-
-@SpringBootApplication
-public class UserService {
-
-        private static final int CAPACITY = 2048;
-        public static void main(String[] args) {
-                SpringApplication application = new SpringApplication(UserService.class);
-                application.setApplicationStartup(new BufferingApplicationStartup(CAPACITY));
-                application.run(args);
-                System.out.println("٩(๑❛ᴗ❛๑)۶  User Service 啟動!!  ٩(๑❛ᴗ❛๑)۶");
-        }
-
-}
