--- user-service/src/main/java/com/userservice/service/HelloService.java
@@ -1,11 +1,0 @@
-package com.userservice.service;
-
-import com.userservice.vo.HelloRequestVO;
-
-public interface HelloService {
-
-    String echo(String input);
-
-    String hello(HelloRequestVO helloRequestVO);
-}
-
