--- user-service/src/main/resources/db/changelog/sqlfile/20240610_1.sql
@@ -1,6 +1,0 @@
-CREATE TABLE IF NOT EXISTS "user".users (
-	id serial4 NOT NULL,
-	"name" varchar NULL,
-	email varchar NULL,
-	CONSTRAINT users_pk PRIMARY KEY (id)
-);
