--- user-service/src/test/java/com/userservice/ApiGatewayTests.java
@@ -1,13 +1,0 @@
-package com.userservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiGatewayTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
